requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
pdfplumber==0.11.4
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==10.4.0
faiss-cpu==1.8.0.post1
transformers==4.44.2
accelerate==0.34.2
sentence-transformers==3.0.1
langchain==0.2.14
langchain-community==0.2.12
langchain-huggingface==0.0.3